+++ extensions/broker-server/src/main/java/de/sovity/edc/ext/brokerserver/BrokerServerExtension.java
@@ -56,6 +56,12 @@ public class BrokerServerExtension implements ServiceExtension {
     @Setting
     public static final String KNOWN_DATASPACE_CONNECTORS = toEdcProp("EDC_BROKER_SERVER_KNOWN_DATASPACE_CONNECTORS");
 
+    @Setting
+    public static final String DELETE_OFFLINE_CONNECTORS_AFTER = toEdcProp("EDC_BROKER_SERVER_DELETE_OFFLINE_CONNECTORS_AFTER");
+
+    @Setting
+    public static final String SCHEDULED_DELETE_OFFLINE_CONNECTORS = toEdcProp("EDC_BROKER_SERVER_SCHEDULED_DELETE_OFFLINE_CONNECTORS");
+
     @Inject
     private ManagementApiConfiguration managementApiConfiguration;