+++ extensions/broker-server/src/main/java/de/sovity/edc/ext/brokerserver/services/config/BrokerServerSettings.java
@@ -29,4 +29,6 @@ public class BrokerServerSettings {
     DataSpaceConfig dataSpaceConfig;
 
     int numThreads;
+
+    Duration deleteOfflineConnectorsAfter;
 }