+++ extensions/broker-server/src/main/java/de/sovity/edc/ext/brokerserver/services/config/BrokerServerSettingsFactory.java
@@ -31,16 +31,18 @@ public class BrokerServerSettingsFactory {
     private final Monitor monitor;
 
     public BrokerServerSettings buildBrokerServerSettings() {
-        var hideOfflineDataOffersAfter = getDurationOrNull(BrokerServerExtension.HIDE_OFFLINE_DATA_OFFERS_AFTER);
+        var hideOfflineDataOffersAfter = getDuration(BrokerServerExtension.HIDE_OFFLINE_DATA_OFFERS_AFTER, null);
         var catalogPagePageSize = config.getInteger(BrokerServerExtension.CATALOG_PAGE_PAGE_SIZE, 20);
         var dataSpaceConfig = buildDataSpaceConfig(config);
         var numThreads = config.getInteger(BrokerServerExtension.NUM_THREADS, 1);
+        var deleteOfflineConnectorsAfter = getDuration(BrokerServerExtension.DELETE_OFFLINE_CONNECTORS_AFTER, Duration.ofDays(5));
 
         return BrokerServerSettings.builder()
                 .hideOfflineDataOffersAfter(hideOfflineDataOffersAfter)
                 .catalogPagePageSize(catalogPagePageSize)
                 .dataSpaceConfig(dataSpaceConfig)
                 .numThreads(numThreads)
+                .deleteOfflineConnectorsAfter(deleteOfflineConnectorsAfter)
                 .build();
     }
 
@@ -76,12 +78,13 @@ public class BrokerServerSettingsFactory {
         return config.getString(BrokerServerExtension.DEFAULT_CONNECTOR_DATASPACE, "Default");
     }
 
-    private Duration getDurationOrNull(@NonNull String configProperty) {
-        var durationAsString = config.getString(configProperty, "");
-        if (StringUtils.isBlank(durationAsString)) {
-            return null;
+    private Duration getDuration(@NonNull String configProperty, Duration defaultValue) {
+        var value = config.getString(configProperty, "");
+
+        if (StringUtils.isBlank(value)) {
+            return defaultValue;
         }
 
-        return Duration.parse(durationAsString);
+        return Duration.parse(value);
     }
 }